+++ examples/saxs.py
@@ -9,8 +9,8 @@
 """
 .. _howto-saxs:
 
-SAXS
-====
+Small-angle X-ray scattering
+============================
 
 Small-angle X-ray scattering (SAXS) can be extracted using MAICoS. To follow this how-to
 guide, you should download the :download:`topology <../../static/water/water.tpr>` and
@@ -22,14 +22,16 @@ First, we import Matplotlib, MDAnalysis, NumPy and MAICoS:
 
 import matplotlib.pyplot as plt
 import MDAnalysis as mda
+from MDAnalysis.analysis.rdf import InterRDF
 
 import maicos
+from maicos.lib.math import compute_form_factor, compute_rdf_structure_factor
 
 
 # %%
-# The `water` system consists of 510 water molecules in the liquid state. The molecules
-# are placed in a periodic cubic cell with an extension of :math:`25 \times 25 \times
-# 25\,\textrm{Å}^3`.
+# The `water` system consists of 510 water molecules in the liquid state. The
+# molecules are placed in a periodic cubic cell with an extent of :math:`25 \times 25
+# \times 25\,\textrm{Å}^3`.
 #
 # Load Simulation Data
 # --------------------
@@ -52,11 +54,12 @@ group_H = u.select_atoms("type H*")
 saxs = maicos.Saxs(u.atoms).run(stop=30)
 
 # %%
-# Note: SAXS computations are extensive calculations. Here, to get an overview of the
-# scattering intensities, we reduce the number of frames to be analyzed from ``101`` to
-# ``30``, by adding the ``stop = 30`` parameter to the ``run`` method. Due to the small
-# number of analyzed frames, the scattering intensities shown in this tutorial should
-# not be used to draw any conclusions from the data.
+# .. Note::
+#   SAXS computations are extensive calculations. Here, to get an overview of the
+#   scattering intensities, we reduce the number of frames to be analyzed from ``101``
+#   to ``30``, by adding the ``stop = 30`` parameter to the ``run`` method. Due to the
+#   small number of analyzed frames, the scattering intensities shown in this tutorial
+#   should not be used to draw any conclusions from the data.
 #
 # Extract the :math:`q` values and the averaged SAXS scattering intensities
 # ``scat_factor`` from the ``results`` attribute:
@@ -70,7 +73,7 @@ scat_factor = saxs.results.scat_factor
 print(scat_factor[:10])
 
 # %%
-# By default, the binwidth in the recipocal :math:`(q)` space is 0.005 1/Å.
+# By default, the binwidth in the recipocal :math:`(q)` space is :math:`0.1 Å^{-1}`.
 #
 # Plot the structure factors profile using:
 
@@ -96,8 +99,8 @@ saxs_O = maicos.Saxs(group_O).run(stop=30)
 saxs_H = maicos.Saxs(group_H).run(stop=30)
 
 # %%
-# Let us plot the results together with the full scattering intensity. Note that here we
-# access the results directly from the ``results`` attribute without storing them in
+# Let us plot the results together with the full scattering intensity. Note that here
+# we access the results directly from the ``results`` attribute without storing them in
 # individual variables before:
 
 fig2, ax2 = plt.subplots()
@@ -111,3 +114,75 @@ ax2.set_ylabel(r"S(q) (arb. units)")
 ax2.legend()
 
 fig2.show()
+
+# %%
+# Connection of the structure factor to the radial distribution function
+# ----------------------------------------------------------------------
+#
+# As in details explained in :ref:`saxs-explanations`, the structure factor can be
+# related to the radial distrubution function (RDF). We denote this structure factor by
+# :math:`S^\mathrm{FT}(q)` since it based on Fourier transforming the RDF. The structure
+# factor which can be directly obtaine from the trajectory is denoted by
+# :math:`S^\mathrm{D}(q)`.
+#
+# To relate these two we first calculate the oxygen-oxygen RDF up to half the box length
+# using :class:`MDAnalysis.analysis.rdf.InterRDF` and save the result in
+# variables for an easier access.
+
+box_lengh = u.dimensions[0]
+
+oo_inter_rdf = InterRDF(
+    g1=group_O, g2=group_O, range=(0, box_lengh / 2), exclude_same="residue"
+).run()
+
+r_oo = oo_inter_rdf.results.bins
+rdf_oo = oo_inter_rdf.results.rdf
+
+# %%
+# We use ``exclude_same="residue"`` to exclude atomic self contributions resulting in a
+# large peak at 0. Next, we convert the RDF into a structure factor using
+# :func:`maicos.lib.math.compute_rdf_structure_factor` and the number density of the
+# oxygens.
+
+density = group_O.n_atoms / u.trajectory.ts.volume
+
+q_rdf, struct_factor_rdf = compute_rdf_structure_factor(
+    rdf=rdf_oo, r=r_oo, density=density
+)
+
+# %%
+# Before we can compare we have to normalize the structure factor from the RDF by the
+# form factor using :func:`maicos.lib.math.compute_form_factor`.
+
+struct_factor_rdf *= compute_form_factor(q_rdf, "O") ** 2
+
+# %%
+# Now we can plot everything together and find that the direct evalation from above and
+# the transformed RDF give the same structure factor.
+
+fig3, ax3 = plt.subplots(2, layout="constrained")
+
+ax3[0].axhline(1, c="gray", ls="dashed")
+ax3[0].plot(r_oo, rdf_oo, label="Oxygen-Oxygen")
+ax3[0].set_xlabel("r / Å")
+ax3[0].set_ylabel("g(r)")
+ax3[0].set_xlim(0, 10)
+
+ax3[1].plot(q_rdf, struct_factor_rdf, label=r"$S^\mathrm{FT}$")
+ax3[1].plot(
+    saxs_O.results.q,
+    saxs_O.results.scat_factor,
+    label=r"$S^\mathrm{D}$",
+    ls="dashed",
+)
+
+ax3[1].set_xlabel("q (1/Å)")
+ax3[1].set_ylabel("S(q) (arb. units)")
+ax3[1].set_xlim(0, 7)
+
+ax3[1].legend()
+ax3[0].legend()
+
+fig3.show()
+
+# %%