+++ src/maicos/lib/math.py
@@ -11,10 +11,11 @@ from typing import Optional, Tuple, Union
 
 import MDAnalysis as mda
 import numpy as np
+from scipy.fftpack import dst
 from scipy.special import ellipe, ellipk
 
 from . import tables
-from ._cutil import compute_structure_factor  # noqa: F401
+from ._cmath import compute_structure_factor  # noqa: F401
 
 
 # Max variation from the mean dt or dk that is allowed (~1e-10 suggested)
@@ -774,3 +775,42 @@ def sa_cylider_intersect_circle(r: float, R: np.ndarray) -> np.ndarray:
         Surface area of the cylinder inside the sphere
     """
     return 8 * R * r * ellipT(np.square(2 * R / r))
+
+
+def compute_rdf_structure_factor(
+    rdf: np.ndarray, r: np.ndarray, density: float
+) -> Tuple[np.ndarray, np.ndarray]:
+    r"""Computes the structure factor based on the radial distribution function (RDF).
+
+    The structure factor :math:`S(q)` based on the RDF :math:`g(r)` is given by
+
+    .. math::
+        S(q) = 1 + 4 \pi \rho \int_0^\infty \mathrm{d}r r
+                         \frac{\sin(qr)}{q} (g(r) - 1)\,
+
+    where :math:`q` is the magnitude of the scattering vector. The calculation is
+    performed via a discrete sine transform as implemented in :func:`scipy.fftpack.dst`.
+
+    For an `example` take a look at :ref:`howto-saxs`.
+
+    Parameters
+    ----------
+    rdf : numpy.ndarray
+        radial distribution function
+    r : numpy.ndarray
+        equally spaced distance array on which rdf is defined
+    density : float
+        number density of particles
+
+    Returns
+    -------
+    q : numpy.ndarray
+        array of q points
+    struct_factor : numpy.ndarray
+        structure factor
+    """
+    dr = r[1] - r[0]
+    q = np.pi / r[-1] * np.arange(1, len(r) + 1)
+    struct_factor = 1 + 4 * np.pi * density * 0.5 * dst((rdf - 1) * r) / q * dr
+
+    return q, struct_factor