+++ src/maicos/modules/saxs.py
@@ -27,15 +27,19 @@ logger = logging.getLogger(__name__)
 class Saxs(AnalysisBase):
     r"""Small angle X-Ray scattering intensities (SAXS).
 
-    The q vectors are binned by their length using a bin_width given by -dq. Using the
-    -nobin option, the raw intensity for each :math:`q_\mathrm{i,j,k}` vector is saved.
-    Note that this only works reliably when using constant box vectors! The possible
-    scattering vectors q can be restricted by a minimal and maximal angle with the
-    z-axis. For 0 and 180, all possible vectors are taken into account. For the
-    scattering factor, the structure factor is multiplied by an atom type specific form
-    factor based on Cromer-Mann parameters. By using the -sel option, atoms can be
-    selected for which the profile is calculated. The selection uses the MDAnalysis
-    selection commands.
+    By default the scattering vectors :math:`\boldsymbol{q}` are binned acording to
+    their length :math:`q` using a bin width given by ``dq``. Setting the option
+    ``bin_spectrum=False``, also the raw scattering vectors and their corresponding
+    Miller indices can be saved. Saving the scattering vectors and Miller indices is
+    only possible when the box vectors are constant in the whole trajectory (NVT) since
+    for changing cells the same Miller indices correspond to different scattering
+    vectors.
+
+    Analyzed scattering vectors :math:`q` can be restricted by a minimal and maximal
+    angle with the z-axis. For ``0`` and ``180``, all possible vectors are taken into
+    account. To obtain the scattering intensities, the structure factor is normalized by
+    an element-specific form factor based on Cromer-Mann parameters
+    :footcite:t:`princeInternationalTablesCrystallography2004`.
 
     For an examples on the usage refer to :ref:`How-to: SAXS<howto-saxs>` and for
     details on the theory see :ref:`saxs-explanations`.
@@ -44,8 +48,9 @@ class Saxs(AnalysisBase):
     ----------
     ${ATOMGROUP_PARAMETER}
     ${BASE_CLASS_PARAMETERS}
-    noboindata : bool
-        Do not bin the data. Only works reliable for NVT!
+    bin_spectrum : bool
+        Bin the spectrum. If :py:obj:`False` Miller indices of q-vector are returned.
+        Only works for NVT simulations.
     startq : float
         Starting q (1/Å)
     endq : float
@@ -64,9 +69,13 @@ class Saxs(AnalysisBase):
     results.q : numpy.ndarray
         length of binned q-vectors
     results.q_indices : numpy.ndarray
-        Miller indices of q-vector (only if noboindata==True)
+        Miller indices of q-vector (only if ``bin_spectrum==False``)
     results.scat_factor : numpy.ndarray
         Scattering intensities
+
+    References
+    ----------
+    .. footbibliography::
     """
 
     def __init__(
@@ -76,10 +85,10 @@ class Saxs(AnalysisBase):
         unwrap: bool = False,
         jitter: float = 0.0,
         concfreq: int = 0,
-        nobin: bool = False,
+        bin_spectrum: bool = True,
         startq: float = 0,
         endq: float = 6,
-        dq: float = 0.005,
+        dq: float = 0.1,
         mintheta: float = 0,
         maxtheta: float = 180,
         output: str = "sq.dat",
@@ -92,7 +101,7 @@ class Saxs(AnalysisBase):
             jitter=jitter,
             concfreq=concfreq,
         )
-        self.nobindata = nobin
+        self.bin_spectrum = bin_spectrum
         self.startq = startq
         self.endq = endq
         self.dq = dq
@@ -104,22 +113,25 @@ class Saxs(AnalysisBase):
         self.mintheta = min(self.mintheta, self.maxtheta)
         self.maxtheta = max(self.mintheta, self.maxtheta)
 
-        if self.mintheta < 0:
-            logger.info(
-                f"mintheta = {self.mintheta}° < 0°: " "Set mininmal angle to 0°."
-            )
-            self.mintheta = 0
-        if self.maxtheta > 180:
-            logger.info(
-                f"maxtheta = {self.maxtheta}° > 180°: " "Set maximal angle to 180°."
-            )
-            self.maxtheta = np.pi
+        if self.mintheta < 0 or self.mintheta > 180:
+            raise ValueError(f"mintheta ({self.mintheta}°) has to between 0 and 180°.")
 
+        if self.maxtheta < 0 or self.maxtheta > 180:
+            raise ValueError(f"maxtheta ({self.maxtheta}°) has to between 0 and 180°.")
+
+        if self.mintheta > self.maxtheta:
+            raise ValueError(
+                f"mintheta ({self.mintheta}°) larger than maxtheta ({self.maxtheta}°)."
+            )
+
+        # Convert angles from degrees to radians
         self.mintheta *= np.pi / 180
         self.maxtheta *= np.pi / 180
 
         self.groups = []
+        self.weights = []
         self.atom_types = []
+
         logger.info("\nMap the following atomtypes:")
         for atom_type in np.unique(self.atomgroup.types).astype(str):
             try:
@@ -132,32 +144,45 @@ class Saxs(AnalysisBase):
                 )
             if element == "DUM":
                 continue
-            self.groups.append(
-                self.atomgroup.select_atoms("type {}*".format(atom_type))
-            )
+
+            group = self.atomgroup.select_atoms("type {}*".format(atom_type))
+
+            self.groups.append(group)
+            # Actual weights (form factors) are applied in post processing after
+            self.weights.append(np.ones(group.n_atoms))
             self.atom_types.append(atom_type)
 
             logger.info("{:>14} --> {:>5}".format(atom_type, element))
 
-        if self.nobindata:
+        if self.bin_spectrum:
+            self.n_bins = int(np.ceil((self.endq - self.startq) / self.dq))
+        else:
             self.box = np.diag(
                 mda.lib.mdamath.triclinic_vectors(self._universe.dimensions)
             )
             self.q_factor = 2 * np.pi / self.box
             self.maxn = np.ceil(self.endq / self.q_factor).astype(int)
-        else:
-            self.n_bins = int(np.ceil((self.endq - self.startq) / self.dq))
 
     def _single_frame(self):
-        # Convert everything to cartesian coordinates.
-        if self.nobindata:
-            self._obs.S_array = np.zeros(list(self.maxn) + [len(self.groups)])
-        else:
-            self._obs.struct_factor = np.zeros([self.n_bins, len(self.groups)])
         box = np.diag(mda.lib.mdamath.triclinic_vectors(self._ts.dimensions))
-        for i, t in enumerate(self.groups):
-            # map coordinates onto cubic cell
-            positions = t.atoms.positions - box * np.round(t.atoms.positions / box)
+
+        if self.bin_spectrum:
+            self._obs.struct_factor = np.zeros([self.n_bins, len(self.groups)])
+        else:
+            if not np.all(box == self.box):
+                raise ValueError(
+                    f"Dimensions in frame {self.frame_index} are different from "
+                    "initial dimenions. Can not use `bin_spectrum=False`."
+                )
+
+            self._obs.S_array = np.zeros(list(self.maxn) + [len(self.groups)])
+
+        for i_group, group in enumerate(self.groups):
+            # Map coordinates onto cubic cell
+            positions = group.atoms.positions - box * np.round(
+                group.atoms.positions / box
+            )
+
             q_ts, S_ts = compute_structure_factor(
                 np.double(positions),
                 np.double(box),
@@ -165,13 +190,12 @@ class Saxs(AnalysisBase):
                 self.endq,
                 self.mintheta,
                 self.maxtheta,
+                self.weights[i_group],
             )
 
-            S_ts *= compute_form_factor(q_ts, self.atom_types[i]) ** 2
+            S_ts *= compute_form_factor(q_ts, self.atom_types[i_group]) ** 2
 
-            if self.nobindata:
-                self._obs.S_array[:, :, :, i] = S_ts
-            else:
+            if self.bin_spectrum:
                 q_ts = q_ts.flatten()
                 S_ts = S_ts.flatten()
                 nonzeros = np.where(S_ts != 0)[0]
@@ -179,35 +203,51 @@ class Saxs(AnalysisBase):
                 q_ts = q_ts[nonzeros]
                 S_ts = S_ts[nonzeros]
 
-                struct_ts = np.histogram(
-                    q_ts, bins=self.n_bins, range=(self.startq, self.endq), weights=S_ts
-                )[0]
-                with np.errstate(divide="ignore", invalid="ignore"):
-                    struct_ts /= np.histogram(
-                        q_ts, bins=self.n_bins, range=(self.startq, self.endq)
-                    )[0]
-                self._obs.struct_factor[:, i] = np.nan_to_num(struct_ts)
+                struct_ts, _ = np.histogram(
+                    a=q_ts,
+                    bins=self.n_bins,
+                    range=(self.startq, self.endq),
+                    weights=S_ts,
+                )
+                bincount, _ = np.histogram(
+                    a=q_ts,
+                    bins=self.n_bins,
+                    range=(self.startq, self.endq),
+                    weights=None,
+                )
+                with np.errstate(invalid="ignore"):
+                    struct_ts /= bincount
+
+                self._obs.struct_factor[:, i_group] = np.nan_to_num(struct_ts)
+            else:
+                self._obs.S_array[:, :, :, i_group] = S_ts
 
     def _conclude(self):
-        if self.nobindata:
-            self.results.scat_factor = self.means.S_array.sum(axis=3)
-            self.results.q_indices = np.array(list(np.ndindex(tuple(self.maxn))))
-            self.results.q = np.linalg.norm(
-                self.results.q_indices * self.q_factor[np.newaxis, :], axis=1
-            )
-        else:
+        if self.bin_spectrum:
             q = np.arange(self.startq, self.endq, self.dq) + 0.5 * self.dq
             nonzeros = np.where(self.means.struct_factor[:, 0] != 0)[0]
             scat_factor = self.means.struct_factor[nonzeros]
 
             self.results.q = q[nonzeros]
             self.results.scat_factor = scat_factor.sum(axis=1)
+        else:
+            self.results.scat_factor = self.means.S_array.sum(axis=3)
+            self.results.q_indices = np.array(list(np.ndindex(tuple(self.maxn))))
+            self.results.q = np.linalg.norm(
+                self.results.q_indices * self.q_factor[np.newaxis, :], axis=1
+            )
 
         self.results.scat_factor /= self.atomgroup.n_atoms
 
     def save(self):
-        """Save the current profiles to a file."""
-        if self.nobindata:
+        """Save results of analysis to file specified by ``output``."""
+        if self.bin_spectrum:
+            self.savetxt(
+                self.output,
+                np.vstack([self.results.q, self.results.scat_factor]).T,
+                columns=["q (1/Å)", "S(q) (arb. units)"],
+            )
+        else:
             out = np.hstack(
                 [
                     self.results.q[:, np.newaxis],
@@ -228,10 +268,4 @@ class Saxs(AnalysisBase):
                 self.output,
                 out,
                 columns=[boxinfo, "q (1/Å)", "q_i", "q_j", "q_k", "S(q) (arb. units)"],
-            )
-        else:
-            self.savetxt(
-                self.output,
-                np.vstack([self.results.q, self.results.scat_factor]).T,
-                columns=["q (1/Å)", "S(q) (arb. units)"],
             )