+++ tests/modules/test_saxs.py
@@ -11,11 +11,14 @@ import os
 import sys
 
 import MDAnalysis as mda
+import numpy as np
 import pytest
-from data import WATER_GRO, WATER_TPR
-from numpy.testing import assert_allclose, assert_almost_equal, assert_equal
+from data import WATER_GRO, WATER_TPR, WATER_TRR
+from MDAnalysis.analysis.rdf import InterRDF
+from numpy.testing import assert_allclose, assert_equal
 
 from maicos import Saxs
+from maicos.lib.math import compute_form_factor, compute_rdf_structure_factor
 
 
 sys.path.append(os.path.join(os.path.dirname(__file__), ".."))
@@ -40,23 +43,64 @@ class TestSaxs(ReferenceAtomGroups):
         u = mda.Universe(WATER_TPR, WATER_GRO)
         return u.atoms
 
+    @pytest.fixture()
+    def ag(self):
+        """Import MDA universe."""
+        u = mda.Universe(WATER_TPR, WATER_TRR)
+        return u.atoms
+
     def test_one_frame(sef, ag_single_frame):
         """Test Saxs on one frame.
 
         Test if the division by the number of frames is correct.
         """
         saxs = Saxs(ag_single_frame, endq=20).run()
-        assert_almost_equal(saxs.results.scat_factor[0], 1.6047, decimal=3)
+        assert_allclose(saxs.results.scat_factor[0], 1.6047, rtol=1e-3)
 
     def test_theta(self, ag_single_frame, tmpdir):
         """Test min & max theta conditions on one frame."""
         with tmpdir.as_cwd():
-            saxs = Saxs(ag_single_frame, mintheta=-10, maxtheta=190).run()
-            saxs.save()
-            assert_allclose(saxs.mintheta, 0)
-            assert_equal(os.path.exists("sq.dat"), True)
+            with pytest.raises(ValueError, match=r"mintheta \(-10°\) has to between 0"):
+                Saxs(ag_single_frame, mintheta=-10, maxtheta=190).run()
 
-    def test_nobindata(self, ag_single_frame):
-        """Test when nobindata is True."""
-        saxs = Saxs(ag_single_frame, nobin=True).run()
+    def test_bin_spectrum(self, ag_single_frame):
+        """Test when bin_spectrum is False."""
+        saxs = Saxs(ag_single_frame, bin_spectrum=False).run()
         assert_equal(type(saxs.q_factor).__name__ == "ndarray", True)
+
+    def test_rdf_comparison(self, ag):
+        """Test if the Fourier transformation of an RDF is the structure factor."""
+        oxy = ag.select_atoms("name OW")
+        L = ag.universe.dimensions[0]  # we have a cubic box
+
+        density = oxy.n_atoms / np.prod(ag.universe.trajectory.ts.volume)
+
+        inter_rdf = InterRDF(
+            oxy,
+            oxy,
+            nbins=300,
+            range=(0, L / 2),
+            exclude_same="residue",
+        ).run()
+
+        q_rdf, struct_factor_rdf = compute_rdf_structure_factor(
+            rdf=inter_rdf.results.rdf,
+            r=inter_rdf.results.bins,
+            density=density,
+        )
+
+        S_fac = Saxs(atomgroup=oxy, dq=0.1).run()
+
+        q = S_fac.results.q
+        scat_factor = S_fac.results.scat_factor
+
+        # Normalize ONLY with respect to the number of particles -> Divide by the form
+        # factor which is applie in the SAXS module
+        struct_factor = scat_factor / compute_form_factor(q, "O") ** 2
+
+        # Interpolate direct method to have same q values. q_rdf covers a larger q
+        # range -> only take those values up the maximum value of 1
+        max_index = sum(q_rdf <= q[-1])
+        struct_factor_interp = np.interp(q_rdf[:max_index], q, struct_factor)
+
+        assert_allclose(struct_factor_interp, struct_factor_rdf[:max_index], atol=3e-2)